修正玩家数据表丢失时 用原始数据恢复

+++ 数据库/techniques_query.c
@@ -23,7 +23,7 @@ extern psocn_dbconn_t conn;
 
 #define TABLE1 CHARACTER_TECHNIQUES
 
-static int db_insert_techniques(uint8_t tech_data[20], uint32_t gc, uint32_t slot) {
+static int db_insert_techniques(techniques_t* tech_data, uint32_t gc, uint32_t slot) {
     memset(myquery, 0, sizeof(myquery));
 
     sprintf(myquery, "INSERT INTO %s "
@@ -39,9 +39,9 @@ static int db_insert_techniques(uint8_t tech_data[20], uint32_t gc, uint32_t slo
         ")",
         TABLE1,
         gc, slot,
-        tech_data[0], tech_data[1], tech_data[2], tech_data[3], tech_data[4], tech_data[5], tech_data[6],
-        tech_data[7], tech_data[8], tech_data[9], tech_data[10], tech_data[11], tech_data[12], tech_data[13],
-        tech_data[14], tech_data[15], tech_data[16], tech_data[17], tech_data[18], tech_data[19]
+        tech_data->all[0], tech_data->all[1], tech_data->all[2], tech_data->all[3], tech_data->all[4], tech_data->all[5], tech_data->all[6],
+        tech_data->all[7], tech_data->all[8], tech_data->all[9], tech_data->all[10], tech_data->all[11], tech_data->all[12], tech_data->all[13],
+        tech_data->all[14], tech_data->all[15], tech_data->all[16], tech_data->all[17], tech_data->all[18], tech_data->all[19]
     );
 
     //DBG_LOG("保存角色更衣室数据 %d", dress_data->create_code);
@@ -57,20 +57,20 @@ static int db_insert_techniques(uint8_t tech_data[20], uint32_t gc, uint32_t slo
     return 0;
 }
 
-static int db_update_techniques(uint8_t tech_data[20], uint32_t gc, uint32_t slot) {
+static int db_update_techniques(techniques_t* tech_data, uint32_t gc, uint32_t slot) {
+
     memset(myquery, 0, sizeof(myquery));
 
     sprintf(myquery, "UPDATE %s SET "
-        "guildcard =  '%" PRIu32 "', slot = '%" PRIu8 "'"
-        ", data0 = '%02X', data1 = '%02X', data2 = '%02X', data3 = '%02X', data4 = '%02X', data5 = '%02X', data6 = '%02X'"
-        ", data7 = '%02X', data8 = '%02X', data9 = '%02X', data10 = '%02X', data11 = '%02X', data12 = '%02X', data13 = '%02X'"
-        ", data14 = '%02X', data15 = '%02X', data16 = '%02X', data17 = '%02X', data18 = '%02X', data19 = '%02X'"
+        "data0 = '%02X', data1 = '%02X', data2 = '%02X', data3 = '%02X', data4 = '%02X', data5 = '%02X', data6 = '%02X', "
+        "data7 = '%02X', data8 = '%02X', data9 = '%02X', data10 = '%02X', data11 = '%02X', data12 = '%02X', data13 = '%02X', "
+        "data14 = '%02X', data15 = '%02X', data16 = '%02X', data17 = '%02X', data18 = '%02X', data19 = '%02X'"
         " WHERE "
         "guildcard = '%" PRIu32 "' AND slot =  '%" PRIu8 "'",
-        TABLE1, gc, slot,
-        tech_data[0], tech_data[1], tech_data[2], tech_data[3], tech_data[4], tech_data[5], tech_data[6],
-        tech_data[7], tech_data[8], tech_data[9], tech_data[10], tech_data[11], tech_data[12], tech_data[13],
-        tech_data[14], tech_data[15], tech_data[16], tech_data[17], tech_data[18], tech_data[19]
+        TABLE1,
+        tech_data->all[0], tech_data->all[1], tech_data->all[2], tech_data->all[3], tech_data->all[4], tech_data->all[5], tech_data->all[6],
+        tech_data->all[7], tech_data->all[8], tech_data->all[9], tech_data->all[10], tech_data->all[11], tech_data->all[12], tech_data->all[13],
+        tech_data->all[14], tech_data->all[15], tech_data->all[16], tech_data->all[17], tech_data->all[18], tech_data->all[19]
         , gc, slot
     );
 
@@ -88,24 +88,45 @@ static int db_update_techniques(uint8_t tech_data[20], uint32_t gc, uint32_t slo
 }
 
 static int db_del_techniques(uint32_t gc, uint32_t slot) {
+    void* result;
+    char** row;
+
     memset(myquery, 0, sizeof(myquery));
 
-    sprintf(myquery, "DELETE FROM %s WHERE guildcard="
+    sprintf(myquery, "SELECT * FROM %s WHERE guildcard="
         "'%" PRIu32 "' AND slot='%" PRIu8 "'", TABLE1, gc,
         slot);
 
-    if (psocn_db_real_query(&conn, myquery)) {
-        SQLERR_LOG("无法清理旧玩家 %s 数据 (GC %"
-            PRIu32 ", 槽位 %" PRIu8 "):\n%s", TABLE1, gc, slot,
-            psocn_db_error(&conn));
-        /* XXXX: 未完成给客户端发送一个错误信息 */
-        return -1;
+    if (!psocn_db_real_query(&conn, myquery)) {
+
+        result = psocn_db_result_use(&conn);
+
+        while ((row = psocn_db_result_fetch(result)) != NULL) {
+
+            sprintf(myquery, "DELETE FROM %s WHERE guildcard="
+                "'%" PRIu32 "' AND slot='%" PRIu8 "'", TABLE1, gc,
+                slot);
+
+            if (psocn_db_real_query(&conn, myquery)) {
+                SQLERR_LOG("无法清理旧玩家 %s 数据 (GC %"
+                    PRIu32 ", 槽位 %" PRIu8 "):\n%s", TABLE1, gc, slot,
+                    psocn_db_error(&conn));
+                /* XXXX: 未完成给客户端发送一个错误信息 */
+                return -1;
+            }
+        }
+
+        psocn_db_result_free(result);
+    }
+    else {
+        SQLERR_LOG("未找到旧玩家 %s 数据 (GC %"
+            PRIu32 ", 槽位 %" PRIu8 ")", TABLE1, gc, slot);
     }
 
     return 0;
 }
 
-int db_update_char_techniques(uint8_t tech_data[20], uint32_t gc, uint32_t slot, uint32_t flag) {
+int db_update_char_techniques(techniques_t* tech_data, uint32_t gc, uint32_t slot, uint32_t flag) {
 
     memset(myquery, 0, sizeof(myquery));
 
@@ -127,7 +148,7 @@ int db_update_char_techniques(uint8_t tech_data[20], uint32_t gc, uint32_t slot,
     return 0;
 }
 
-int db_get_char_techniques(uint32_t gc, uint8_t slot, uint8_t tech_data[20], int check) {
+int db_get_char_techniques(uint32_t gc, uint8_t slot, techniques_t* tech_data, int check) {
     void* result;
     char** row;
 
@@ -162,7 +183,7 @@ int db_get_char_techniques(uint32_t gc, uint8_t slot, uint8_t tech_data[20], int
     if (row != NULL) {
         int i2 = 2;
         for (int i = 0; i < MAX_TECH_LEVEL; i++) {
-            tech_data[i] = (uint8_t)strtoul(row[i2], NULL, 16);
+            tech_data->all[i] = (uint8_t)strtoul(row[i2], NULL, 16);
             i2++;
         }
     }
